test(query-builder): move join regression tests into query-builder/join/
