+++ test/functional/query-builder/join/custom-on-condition/entity/PostReview.ts
@@ -3,7 +3,7 @@ import {
     PrimaryGeneratedColumn,
     Column,
     ManyToOne,
-} from "../../../../src"
+} from "../../../../../../src"
 import { Post } from "./Post"
 
 @Entity()
