+++ test/functional/query-builder/join/escape-keyword-aliases/entity/Group.ts
@@ -1,0 +1,12 @@
+import { Entity } from "../../../../../../src/decorator/entity/Entity"
+import { Column } from "../../../../../../src/decorator/columns/Column"
+import { PrimaryGeneratedColumn } from "../../../../../../src/decorator/columns/PrimaryGeneratedColumn"
+
+@Entity()
+export class Group {
+    @PrimaryGeneratedColumn()
+    id: number
+
+    @Column()
+    name: string
+}
