+++ test/functional/query-builder/join/on-clause-property-name-regex/entity/Abbreviation.ts
@@ -1,0 +1,12 @@
+import { Entity } from "../../../../../../src/decorator/entity/Entity"
+import { PrimaryGeneratedColumn } from "../../../../../../src/decorator/columns/PrimaryGeneratedColumn"
+import { Column } from "../../../../../../src/decorator/columns/Column"
+
+@Entity()
+export class Abbreviation {
+    @PrimaryGeneratedColumn()
+    id: number
+
+    @Column()
+    name: string
+}
