+++ test/functional/query-builder/join/on-clause-property-name-regex/entity/Post.ts
@@ -1,9 +1,9 @@
-import { Entity } from "../../../../src/decorator/entity/Entity"
-import { PrimaryGeneratedColumn } from "../../../../src/decorator/columns/PrimaryGeneratedColumn"
+import { Entity } from "../../../../../../src/decorator/entity/Entity"
+import { PrimaryGeneratedColumn } from "../../../../../../src/decorator/columns/PrimaryGeneratedColumn"
 import { Author } from "./Author"
 import { Abbreviation } from "./Abbreviation"
-import { OneToOne } from "../../../../src/decorator/relations/OneToOne"
-import { JoinColumn } from "../../../../src/decorator/relations/JoinColumn"
+import { OneToOne } from "../../../../../../src/decorator/relations/OneToOne"
+import { JoinColumn } from "../../../../../../src/decorator/relations/JoinColumn"
 
 @Entity()
 export class Post {
