--- test/github-issues/401/entity/Group.ts
@@ -1,12 +1,0 @@
-import { Entity } from "../../../../src/decorator/entity/Entity"
-import { Column } from "../../../../src/decorator/columns/Column"
-import { PrimaryGeneratedColumn } from "../../../../src/decorator/columns/PrimaryGeneratedColumn"
-
-@Entity()
-export class Group {
-    @PrimaryGeneratedColumn()
-    id: number
-
-    @Column()
-    name: string
-}
