--- test/other-issues/join-empty-relations/entity/Category.ts
@@ -1,19 +1,0 @@
-import { Entity } from "../../../../src/decorator/entity/Entity"
-import { PrimaryGeneratedColumn } from "../../../../src/decorator/columns/PrimaryGeneratedColumn"
-import { Column } from "../../../../src/decorator/columns/Column"
-import { JoinTable } from "../../../../src/decorator/relations/JoinTable"
-import { ManyToMany } from "../../../../src/decorator/relations/ManyToMany"
-import { Author } from "./Author"
-
-@Entity()
-export class Category {
-    @PrimaryGeneratedColumn()
-    id: number
-
-    @Column()
-    name: string
-
-    @ManyToMany(() => Author)
-    @JoinTable()
-    authors: Author[]
-}
