--- test/other-issues/join-empty-relations/entity/Post.ts
@@ -1,19 +1,0 @@
-import { Entity } from "../../../../src/decorator/entity/Entity"
-import { PrimaryGeneratedColumn } from "../../../../src/decorator/columns/PrimaryGeneratedColumn"
-import { Column } from "../../../../src/decorator/columns/Column"
-import { Category } from "./Category"
-import { ManyToMany } from "../../../../src/decorator/relations/ManyToMany"
-import { JoinTable } from "../../../../src/decorator/relations/JoinTable"
-
-@Entity()
-export class Post {
-    @PrimaryGeneratedColumn()
-    id: number
-
-    @Column()
-    title: string
-
-    @ManyToMany(() => Category)
-    @JoinTable()
-    categories: Category[]
-}
